+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 * differences from [previous tag](/../../compare/v1.7.1…main).
+* Added [version-related unit tests](./tests/test_bme_version.sh).
 
 ## v1.7.1 (2024-FEB-04)
 * differences from [previous tag](/../../compare/v1.7.0…v1.7.1).